Change from a staticly compiled node.js pipeline to an AJAX dynamically loaded pipeline
The biggest advantage being removing the requirement for a node.js pipeline. The disadvantages with this is the removement of node.js tools and the requirement of the dataset to to be online, but seeing as repository is hosted on github anyway it's not really a big issue in this case.